+++ packages/api/lib/testUtils.js
@@ -81,6 +81,7 @@ function fakeGranuleFactory(status = 'completed') {
   return {
     granuleId: randomId('granule'),
     collectionId: constructCollectionId('fakeCollection', 'v1'),
+    producerGranuleId: randomId('producerGranuleId'),
     status,
     execution: getExecutionUrlFromArn(randomId('execution')),
     createdAt: Date.now(),