+++ packages/db/src/migrations/20250425134823_granules_add_producer_granule_id.ts
@@ -2,7 +2,6 @@ import { Knex } from 'knex';
 
 export const up = async (knex: Knex): Promise<void> => {
   if (!await knex.schema.hasColumn('granules', 'producer_granule_id')) {
-
     await knex.schema.table('granules', (table) => {
       table
         .string('producer_granule_id')
@@ -14,6 +13,8 @@ export const up = async (knex: Knex): Promise<void> => {
     await knex.schema.table('granules', (table) => {
       table.string('producer_granule_id').notNullable().alter();
     });
+
+    await knex.raw('CREATE INDEX CONCURRENTLY IF NOT EXISTS granules_producer_granule_id_index ON granules(producer_granule_id');
   }
 };